fixed bad checkin

+++ UN.CYBERCOM.Contracts/CYBERCOM/ContractDefinition/VotingParameters.cs
@@ -19,14 +19,14 @@ namespace UN.CYBERCOM.Contracts.CYBERCOM.ContractDefinition
         public virtual BigInteger OutputCountForGroup { get; set; }
         [Parameter("uint256", "outputCountForMember", 4)]
         public virtual BigInteger OutputCountForMember { get; set; }
-        [Parameter("int32", "voteDenominator", 5)]
-        public virtual int VoteDenominator { get; set; }
-        [Parameter("int32", "voteNumerator", 6)]
-        public virtual int VoteNumerator { get; set; }
-        [Parameter("int32", "sumDenominator", 7)]
-        public virtual int SumDenominator { get; set; }
-        [Parameter("int32", "sumNumerator", 8)]
-        public virtual int SumNumerator { get; set; }
+        [Parameter("uint256", "voteDenominator", 5)]
+        public virtual BigInteger VoteDenominator { get; set; }
+        [Parameter("uint256", "voteNumerator", 6)]
+        public virtual BigInteger VoteNumerator { get; set; }
+        [Parameter("uint256", "sumDenominator", 7)]
+        public virtual BigInteger SumDenominator { get; set; }
+        [Parameter("uint256", "sumNumerator", 8)]
+        public virtual BigInteger SumNumerator { get; set; }
         [Parameter("bool", "avgVotes", 9)]
         public virtual bool AvgVotes { get; set; }
     }
